use version from release. otherwise fall back to git commit hash
When publishing, we always define a tag. This is used as version number
for creating and publishing the PyPI package. When manually triggering
the publish workflow, we don't have a release version (but we will also
not publish to PyPI), so we simply use the Git commit hash as the version
number (fallback done by setup.py internally).